--- src/main/java/com/microservices/choreographyorchestrator/ChoreographyOrchestratorApplication.java
@@ -1,13 +1,0 @@
-package com.microservices.choreographyorchestrator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ChoreographyOrchestratorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChoreographyOrchestratorApplication.class, args);
-	}
-
-}
