--- src/main/java/com/microservices/choreographyorchestrator/domain/definition/CompensationInfo.java
@@ -1,10 +1,0 @@
-package com.microservices.choreographyorchestrator.domain.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CompensationInfo {
-    private KafkaCompensation kafkaCompensation;
-}
