--- src/main/java/com/microservices/choreographyorchestrator/domain/definition/ExecutionInfo.java
@@ -1,12 +1,0 @@
-package com.microservices.choreographyorchestrator.domain.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ExecutionInfo {
-    private String channelName;
-
-    private String eventType;
-}
