--- src/main/java/com/microservices/choreographyorchestrator/domain/definition/KafkaRetry.java
@@ -1,12 +1,0 @@
-package com.microservices.choreographyorchestrator.domain.definition;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class KafkaRetry {
-    private String topicName;
-}
