--- src/main/java/com/microservices/choreographyorchestrator/dto/definition/CompensationInfoDto.java
@@ -1,10 +1,0 @@
-package com.microservices.choreographyorchestrator.dto.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CompensationInfoDto {
-    private KafkaCompensationInfoDto kafkaCompensationInfo;
-}
