--- src/main/java/com/microservices/choreographyorchestrator/dto/definition/KafkaCompensationInfoDto.java
@@ -1,10 +1,0 @@
-package com.microservices.choreographyorchestrator.dto.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class KafkaCompensationInfoDto {
-    private String topicName;
-}
