--- src/main/java/com/microservices/choreographyorchestrator/dto/definition/KafkaRetryDto.java
@@ -1,10 +1,0 @@
-package com.microservices.choreographyorchestrator.dto.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class KafkaRetryDto {
-    private String topicName;
-}
