--- src/main/java/com/microservices/choreographyorchestrator/dto/definition/SuccessExecutionInfoDto.java
@@ -1,12 +1,0 @@
-package com.microservices.choreographyorchestrator.dto.definition;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SuccessExecutionInfoDto {
-    private String channelName;
-
-    private String eventType;
-}
