--- src/main/java/com/microservices/choreographyorchestrator/repository/SagaStepTransitionRepository.java
@@ -1,15 +1,0 @@
-package com.microservices.choreographyorchestrator.repository;
-
-import com.microservices.choreographyorchestrator.domain.SagaStepTransitionEvent;
-import com.microservices.choreographyorchestrator.domain.definition.SagaStepDefinition;
-import org.springframework.data.neo4j.repository.Neo4jRepository;
-
-import java.util.List;
-
-public interface SagaStepTransitionRepository extends Neo4jRepository<SagaStepTransitionEvent, Long> {
-    List<SagaStepTransitionEvent> findSagaStepTransitionEventsByPreviousStep(SagaStepDefinition previousStep);
-
-    List<SagaStepTransitionEvent> findSagaStepTransitionEventsByNextStep(SagaStepDefinition nextStep);
-
-    List<SagaStepTransitionEvent> findSagaStepTransitionEventsByPreviousStepAndNextStep(SagaStepDefinition previousStep, SagaStepDefinition nextStep);
-}
