+++ src/main/java/com/microservices/saga/choreography/supervisor/domain/definition/CompensationInfo.java
@@ -1,0 +1,10 @@
+package com.microservices.saga.choreography.supervisor.domain.definition;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CompensationInfo {
+    private KafkaCompensation kafkaCompensation;
+}
