+++ src/main/java/com/microservices/saga/choreography/supervisor/domain/definition/ExecutionInfo.java
@@ -1,0 +1,12 @@
+package com.microservices.saga.choreography.supervisor.domain.definition;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ExecutionInfo {
+    private String channelName;
+
+    private String eventType;
+}
