+++ src/main/java/com/microservices/saga/choreography/supervisor/domain/definition/KafkaCompensation.java
@@ -1,0 +1,12 @@
+package com.microservices.saga.choreography.supervisor.domain.definition;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class KafkaCompensation {
+    private String topicName;
+}
