+++ src/main/java/com/microservices/saga/choreography/supervisor/domain/definition/SagaStepDefinition.java
@@ -1,0 +1,30 @@
+package com.microservices.saga.choreography.supervisor.domain.definition;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.neo4j.ogm.annotation.GeneratedValue;
+import org.neo4j.ogm.annotation.Id;
+import org.neo4j.ogm.annotation.NodeEntity;
+
+@NodeEntity
+@Getter
+@Setter
+public class SagaStepDefinition {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String sagaName;
+
+    private String stepName;
+
+    private ExecutionInfo successExecutionInfo;
+
+    private ExecutionInfo failExecutionInfo;
+
+    private CompensationInfo compensationInfo;
+
+    private RetryInfo retryInfo;
+}
+
+
