+++ src/main/java/com/microservices/saga/choreography/supervisor/dto/EventDto.java
@@ -1,0 +1,18 @@
+package com.microservices.saga.choreography.supervisor.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class EventDto {
+    private Long eventId;
+
+    private Long sagaInstanceId;
+
+    private String eventName;
+
+    private String sagaName;
+}
