+++ src/main/java/com/microservices/saga/choreography/supervisor/dto/definition/KafkaCompensationInfoDto.java
@@ -1,0 +1,10 @@
+package com.microservices.saga.choreography.supervisor.dto.definition;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class KafkaCompensationInfoDto {
+    private String topicName;
+}
