+++ src/main/java/com/microservices/saga/choreography/supervisor/dto/definition/SuccessExecutionInfoDto.java
@@ -1,0 +1,12 @@
+package com.microservices.saga.choreography.supervisor.dto.definition;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SuccessExecutionInfoDto {
+    private String channelName;
+
+    private String eventType;
+}
