--- src/test/java/com/microservices/choreographyorchestrator/ChoreographyOrchestratorApplicationTests.java
@@ -1,13 +1,0 @@
-package com.microservices.choreographyorchestrator;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChoreographyOrchestratorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
