+++ src/test/java/com/microservices/saga/choreography/supervisor/ChoreographyOrchestratorApplicationTests.java
@@ -1,0 +1,13 @@
+package com.microservices.saga.choreography.supervisor;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ChoreographyOrchestratorApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
